requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
pydantic-settings==2.7.0
httpx==0.28.0
beautifulsoup4==4.12.3
reportlab==4.2.5
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
